my_plugins/nat_anthropic/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nat_anthropic"
version = "0.1.0"
description = "Anthropic Claude LLM provider for NeMo Agent Toolkit"
requires-python = ">=3.11"
dependencies = [
    "nvidia-nat>=1.3.0",
    "anthropic>=0.40.0",
]

[project.entry-points."nat.llm_provider"]
anthropic = "nat_anthropic.anthropic_llm:anthropic_llm"

[tool.setuptools.packages.find]
where = ["."]
